remove user authentication
