+++ client/src/models/reports.ts
@@ -1,6 +1,5 @@
 export interface Reports {
   reportNo: string;
-  userId: string;
   _id: string;
   name: string;
   work: string;
