--- client/src/pages/Home.tsx
@@ -1,11 +1,0 @@
-import { LoginPage } from "./LoginPage";
-import { LogoutPage } from "./LogoutPage";
-import { User } from "../models/user";
-
-interface HomePageProps {
-  logInUser: User | null;
-}
-
-export const HomePage = ({ logInUser }: HomePageProps) => {
-  return <>{logInUser ? <LoginPage /> : <LogoutPage />}</>;
-};
