--- client/src/pages/LogoutPage.tsx
@@ -1,7 +1,0 @@
-export const LogoutPage = () => {
-  return (
-    <div className="flex flex-col items-center">
-      <p>Please login to see your reports</p>
-    </div>
-  );
-};
