+++ server/src/controllers/reportController.ts
@@ -6,7 +6,7 @@ export const getAllReports = async (
   res: Response
 ): Promise<void> => {
   try {
-    const reports = await reportData.find({ userId: req.session.userId });
+    const reports = await reportData.find({});
     res.status(200).json({ reports });
   } catch (error) {
     res.status(500).json({ message: error });
@@ -16,10 +16,8 @@ export const getAllReports = async (
 export const getReport = async (req: Request, res: Response): Promise<void> => {
   try {
     const { id } = req.params;
-    const userId = req.session.userId;
     const report = await reportData.findOne({
       _id: id,
-      userId,
     });
     res.status(200).json(report);
   } catch (error) {
@@ -32,17 +30,11 @@ export const createReport = async (
   res: Response
 ): Promise<void> => {
   try {
-    const userId = req.session.userId; // retrieve userId from session
-    if (!userId) {
-      res.status(401).json({ message: "User not authenticated" });
-      return;
-    }
     const report = await reportData.create({
       name: req.body.name,
       work: req.body.work,
       actions: req.body.actions,
       date: req.body.date,
-      userId: userId,
     });
     res.status(200).json(report);
   } catch (error) {
@@ -56,15 +48,10 @@ export const updateReport = async (
 ): Promise<void> => {
   try {
     const { id } = req.params;
-    const userId = req.session.userId;
-    const report = await reportData.findOneAndUpdate(
-      { _id: id, userId },
-      req.body,
-      {
-        new: true,
-        runValidators: true,
-      }
-    );
+    const report = await reportData.findOneAndUpdate({ _id: id }, req.body, {
+      new: true,
+      runValidators: true,
+    });
     res.status(200).json(report);
   } catch (error) {
     res.status(500).json({ message: error });
